requirements.txt
astroid==1.3.8
braintree==3.20.0
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-braces==1.8.1
django-debug-toolbar==1.3.2
django-extensions==1.5.6
django-extra-views==0.7.1
django-grappelli==2.7.1
django-jsignature==0.8
django-model-utils==2.3.1
django-widget-tweaks==1.4.1
docopt==0.6.2
docutils==0.12
factory-boy==2.5.2
fake-factory==0.5.2
gunicorn==19.3.0
html-linter==0.1.9
logilab-common==1.0.2
nose==1.3.7
Pillow==2.9.0
psycopg2==2.6.1
pylint==1.4.4
pystache==0.5.4
python-gflags==2.0
pytz==2015.4
requests==2.7.0
six==1.9.0
sqlparse==0.1.16
static3==0.6.1
template-remover==0.1.9
termcolor==1.1.0
Unipath==1.1
wheel==0.24.0
